+++ backend/FetchResultsUtils.js
